devel/cl-asdf: [WIP] Update to 3.3.6
Summary:
This update also fixes the problem that SBCL would disable the port ASDF
in favor of its older, bundled ASDF.  SBCL did this because the port was not
installing version.lisp-expr.

This is still a work-in-progress.  A problem that still needs to be addressed is
that this newer ASDF wants to auto-generate fasl files under
${LOCALBASE}/lib/common-lisp/asdf/build/ after it is installed.  A solution is
to create a devel/cl-asdf-sbcl port, like our other CL ports.  However, this
newer ASDF is also writing under ${LOCALBASE}/lib/common-lisp/asdf/build/ when
building the fasl package.  I hope to tackle this problem in the next few days,
but put this here in case anyone has suggestions.

A suggestion from https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=243176, if I
understand correctly, is to eliminate the *-sbcl ports and presumably eliminate
(or heavily modify) ${PORTSDIR}/devel/cl-asdf/bsd.cl-asdf.mk and
${PORTDIR}/devel/cl-asdf/files/asdf-init.lisp and allow SBCL to write the fasl
files under the user's home directory.

PR:		243176
Sponsored by:	The FreeBSD Foundation

Reviewers: olgeni, ashish

Subscribers: mat

Differential Revision: https://reviews.freebsd.org/D37634