cpu-microcode-intel: Add knob for cpu-microcode-rc dependency
The RC script included in sysutils/cpu-microcode-rc is one of two
methods for updating an Intel CPU microcode.  Give users the option of
depending on that script. If the dependency is unwanted, then do not
install the split microcode files, which are only required when using
cpucontrol(8), which is called in the RC script.